+++ examples/create-client.ts
@@ -4,24 +4,27 @@ import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
 (async () => {
     try {
         const flex = new Flex(EXAMPLES_FLEX_CONFIG);
+        const traderId = CONFIG.trader.id;
         const clientAddress = await Client.deploy(flex, {
             everWallet: {
-                address: "0:b4da2773b3566c8799ff8292bb1058662d143556a7ac8a129c481a38657cbd33",
-                signer: "msig",
+                address: "0:d807caf6df3a7c2bb0b64915613eca9d8f17ca1de0b938dfdcbb9b4ff30c4526",
+                signer: "everWallet",
             },
-            signer: "flex-client-1",
+            signer: "everWallet",
+            
         });
+        console.log(`Client: ${clientAddress}}`);
+        
         await Trader.deploy(flex, {
             client: {
                 address: clientAddress,
-                signer: "flex-client-2",
+                signer: "everWallet",
             },
-            id: CONFIG.trader.id,
-            name: "Trader 1",
-            pubkey: await flex.evr.signers.resolvePublicKey(CONFIG.trader.signer),
+            id: traderId,
+            name: "trader_1",
+            pubkey: "162c6c708018da073729dd4a60118425dd917e44653383f1faed4d16b94af30b" // Trader's pubkey
         });
 
-        console.log(`Client: ${clientAddress}}`);
         await flex.close();
     } catch (err) {
         console.error(err);