+++ flex/trader/deploy-ever-wallet.ts
@@ -1,5 +1,5 @@
 import { Flex } from "../flex";
-import { WrapperAccount } from "../../contracts";
+import { WrapperEverAccount } from "../../contracts";
 import { AccountOptionsEx } from "../../contracts/account-ex";
 import { EverWallet, toUnits, uint256 } from "../web3";
 
@@ -26,12 +26,12 @@ export type DeployTraderEverWalletOptions = {
     tokens: number,
 
     /**
-     * TODO: Total fees for deposit
+     * Amount of native EVERs that the deposit message carries. Later on, DEX wallet will spend them to pay for gas. 
      */
     evers?: number,
     /**
      * Minimum amount of EVERs on DEX wallet. If balance drops below this amount,
-     * wallet is topped-up from Trader's Index wallet.
+     * wallet is automatically topped-up from the Trader's Index wallet.
      */
     keepEvers?: number,
 }
@@ -41,6 +41,7 @@ export type EverWalletInfo = {
 }
 
 const DEFAULTS = {
+    
     evers: 15,
     keepEvers: 12,
 };
@@ -51,7 +52,7 @@ export async function deployTraderEverWallet(
     options: DeployTraderEverWalletOptions,
 ): Promise<EverWalletInfo> {
     const pubkey = uint256(options.traderId);
-    const wrapper = await flex.evr.accounts.get(WrapperAccount, options.wrapperAddress);
+    const wrapper = await flex.evr.accounts.get(WrapperEverAccount, options.wrapperAddress);
     const walletAddress = (await wrapper.getWalletAddress({
         owner: options.clientAddress,
         pubkey,
@@ -63,7 +64,7 @@ export async function deployTraderEverWallet(
         dest: options.wrapperAddress,
         value: toUnits(options.tokens + evers),
         payload: {
-            abi: WrapperAccount.package.abi,
+            abi: WrapperEverAccount.package.abi,
             fn: "onEverTransfer",
             params: {
                 tokens: toUnits(options.tokens),
@@ -77,7 +78,7 @@ export async function deployTraderEverWallet(
             },
         },
     });
-    await everWallet.topUp(walletAddress, evers);
+    
     return {
         address: walletAddress,
     };